#100 [U-ways] Decouple class loader extension from Gradle domain by setting extension receiver to be List<File> instead of org.gradle.api.file.FileCollection
